Leo Romero - Refactored Search component - Delted trash icon - Added react memo in GifCard to optimized at performance - Changed Layout styles and component - Renamed Header component to Navbar component - Refactored styles in Navbar component - Deleted Paginator component - Refactored Search component to optimized at performance and clean code - Refactored SearchPage
