+++ src/components/Layout/index.js
@@ -1,11 +1,13 @@
-import Header from 'components/Header';
+import Navbar from 'components/Navbar';
 import './style.css';
 
 export default function Layout({ children, showSearcher }) {
   return (
     <div className="main-container">
-      <Header showSearcher={showSearcher} />
-      {children}
+      <Navbar showSearcher={showSearcher} />
+      <div className="container-children">
+        {children}
+      </div>
     </div>
   );
 }
