+++ src/components/Layout/style.css
@@ -3,8 +3,13 @@
   align-items: center;
   flex-direction: column;
   gap: 32px;
-  padding: 32px 8px;
-  padding-bottom: 140px;
+  padding: 0 8px 140px 8px;
+  width: 100%;
+}
+
+.container-children {
+  display: flex;
+  flex-direction: column;
   width: 100%;
 }
 
@@ -12,6 +17,11 @@
   .main-container {
     gap: 48px;
     max-width: 1440px;
-    padding: 40px 20px;
+    padding: 0 20px 40px 20px;
+  }
+
+  .container-children {
+    padding-top: 160px;
+    gap: 32px;
   }
 }
