+++ src/components/Navbar/style.css
@@ -60,6 +60,14 @@
 }
 
 @media (min-width: 1024px) {
+  .header {
+    padding-top: 40px;
+    background-color: var(--color-primary-dark);
+    box-shadow: -3px 8px 18px 0px var(--color-primary-dark);
+    position: fixed;
+    z-index: 40;
+  }
+
   .header__github--link {
     padding: 12px 32px;
   }
