--- src/components/Paginator/style.css
@@ -1,24 +1,0 @@
-.paginator {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.paginator__button {
-  background-color: var(--color-primary-700);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 12px;
-  border-radius: 100px;
-  padding: 8px 16px;
-  transition: all ease-in-out .2s;
-}
-
-.paginator__button:hover {
-  background-color: var(--color-primary-900);
-}
-
-.paginator__text {
-  color: var(--color-primary-200);
-}
